pyproject.toml
[tool.poetry]
name = "butler"
version = "0.1.0"
description = ""
authors = ["wz <wz@waifoo.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "==4.2.*"
djangorestframework = "^3.14.0"
ipython = "^8.13.2"
django-extensions = "^3.2.1"
environs = {extras = ["django"], version = "^9.5.0"}
django-cors-headers = "^4.0.0"
psycopg = {extras = ["binary"], version = "^3.1.9"}
daphne = "^4.0.0"
django-ipware = "^5.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
